# contracts/openapi-v1.baseline.json
{
  "components": {
    "schemas": {
...
      "UpdateIntegrationRequest": {
        "additionalProperties": false,
        "properties": {
          "apiKey": {
            "type": "string"
          },
          "apiSecret": {
            "type": [
              "null",
              "string"
            ]
          },
          "baseUrl": {
            "type": "string"
          },
          "createSuppliersInErp": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "defaultSupplierGroup": {
            "type": [
              "null",
              "string"
            ]
          },
          "isEnabled": {
            "type": "boolean"
          }
        },
        "required": [
          "baseUrl",
          "apiKey",
          "apiSecret",
          "isEnabled"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MotsSupplierPortal.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/admin/integrations": {
      "get": {
        "description": "Requires the `admin.integrations.manage` permission.",
        "operationId": "ListIntegrations",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "FORBIDDEN"
          }
        },
        "tags": [
          "Integrations"
        ],
        "x-required-permissions": [
          "admin.integrations.manage"
        ]
      }
    },
    "/api/v1/admin/integrations/{key}": {
      "put": {
        "description": "Requires the `admin.integrations.manage` permission.",
        "operationId": "UpdateIntegration",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIntegrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "FORBIDDEN"
          }
        },
        "tags": [
          "Integrations"
        ],
        "x-required-permissions": [
          "admin.integrations.manage"
        ]
      }
    },
    "/api/v1/admin/integrations/{key}/supplier-groups": {
      "get": {
        "description": "Requires the `admin.integrations.manage` permission.",
        "operationId": "ListErpSupplierGroups",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "FORBIDDEN"
          }
        },
        "tags": [
          "Integrations"
        ],
        "x-required-permissions": [
          "admin.integrations.manage"
        ]
      }
    },
    "/api/v1/admin/integrations/{key}/test": {
      "post": {
        "description": "Requires the `admin.integrations.manage` permission.",
        "operationId": "TestIntegration",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "FORBIDDEN"
          }
        },
        "tags": [
          "Integrations"
        ],
        "x-required-permissions": [
          "admin.integrations.manage"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
...
}